static/DeltaAI/requirements.txt
Flask==2.3.3
google-cloud-vision==3.4.4
Werkzeug==2.3.7
Pillow==10.0.1
